don't memoize CartesianVector magnitude
It just occurred to me that this strategy wasn't thread-safe.

I might consider making this a lazy val, but perhaps memory usage
is a more important concern.

If a user who is going to be referring to magnitude repeatedly could
convert to polar coordinates. Also, I could add another Vector type
that contains both XY and polar coordinates, so users who know that
they'll want to refer to both could convert to that.