remove BigDecimal support
Decided it just never makes sense to use BigDecimal in this context. It
especially gains nothing when angles come into play, because most trig
functions are going to be dealing with irrational values anyway.

Switching the test from BigDecimal to Double uncovered an error.
Dividing a BigDecimal by 0 threw ArithmeticException, but dividing a
Double by 0 returns Infinity. I don't know any decent way around this,
so I'm just going to drop that test and probably forget about it and
hope it's never an issue.