give up on Scalar type class, use Double
Trying to be generic over Float and Double was, more trouble than it
was worth. I found myself left with two similarly unappealing options:

1) Keep the context bound for S at the class level. This is not too
   bad syntactically, but it results in an extra field on every class
   that is parameterized on S.

2) Move the context bound to the method level. This adds a lot of
   boilerplate; every method on Vector would need to declare that
   implicit parameter and import from it. This code explosion would
   likely have continued into most other classes as well.

Because *every* class and method in which S is involved needs the
Scalar type class, it just got too cumbersome to deal with.

This is quite disappointing, and I'm starting to feel like Scala has to
some extent failed in its goal of enabling code that is both generic and
efficient.