Rename `Vector` to `Point`
I think "point" is a more appropriate name because it's more in tune
with the geometric semantics, on theme with the library name "space",
and less about the algebra. The subtypes remain called CartesianVector
and PolarVector, because "vector" is fitting at that level of
abstraction where the pairs of numbers are materialized.