Add Square, stub Rhombus
Also includes some refactor to Rectangle.

Add OrthogonalBoundingBox

Replace OrthogonalRectangle sizeX,sizeY:Double fields with one
size:CartesianVector field. I'm on the fence about whether that
constitutes an abuse of the type. It might be slightly misleading, even
though it's perhaps appropriate to refer to a width/height as a
cartesian vector, it's less appropriate to consider it as a Point. But
the result reads better; `size.x` looks nicer than `sizeX`.